An attempt to org-mode the MM statistic block
One problem I had during play was I kept having to refer back to the MM.
That meant I was giving up screen real estate for the PDF instead of
using it for emacs. Putting the stat block in org-mode may help me to
refer to specs faster? The idea would be to look up the stat in MM and
then (quickly) copy/paste it to org-mode for use in the encounter.